feat: allow editing of individual class details
